refactor(mcp): rename weacpx-mcp-*.ts files to xacpx-mcp-*.ts
